Add Ansible Molecule environment
Added a Makefile to create the Python3 virtualenv environment to set up
Ansible molecule.

Molecule was then initialized with:

  molecule init scenario -r ansible-role-samba --driver-name docker

As we only work on Debian servers, we currently only use Debian docker
images as molecule platform.

I renamed the 'default' scenario to 'member' to test the new
'samba_role: member' feature. For now, I created a 'default' symlink to
the 'member', so 'molecule test' can be run without specifying a
scenario.

The 'default' scenario will test the role with the default variables.

Added 'namespace' in meta/main.yml to fix the molecule error:

  ERROR    Computed fully qualified role name of samba does not follow current galaxy requirements.
  Please edit meta/main.yml and assure we can correctly determine full role name:

  galaxy_info:
  role_name: my_name  # if absent directory name hosting role is used instead
  namespace: my_galaxy_namespace  # if absent, author is used instead

  Namespace: https://galaxy.ansible.com/docs/contributing/namespaces.html#galaxy-namespace-limitations
  Role: https://galaxy.ansible.com/docs/contributing/creating_role.html#role-names

  As an alternative, you can add 'role-name' to either skip_list or warn_list.

The GitHub handle of the original Author is @lusico. I used that as
namespace.

If we do not set the namespace, the github-workflow molecule fails,
because it can not find the ansible role in the given directory.

The role has then to be called 'luisico.samba' (in converge.yml).kgg